Clarify 404 error for unknown API resources
