+++ functions/api/index.js
@@ -2,6 +2,7 @@
 import { connectDB } from '../db/connection.js';
 import { verifyTelegramWebAppData } from '../utils/auth.js';
 import { handleChannelsApi } from './channels.js';
+import { handleTicketApi } from './ticket.js';
 
 export async function handleApiRequest(request, env, path) {
   const db = await connectDB(env);
@@ -52,13 +53,13 @@ export async function handleApiRequest(request, env, path) {
       
       case 'ticket':
         return await handleTicketApi(request, env, params, user);
-      
+
       case 'channels':
-        return await handleChannelsApi(request, db, params, user);
+        return await handleChannelsApi(request, db, params, user, env);
       
       default:
         return {
-          body: JSON.stringify({ error: 'Not found' }),
+          body: JSON.stringify({ error: `Unknown API resource: ${resource}` }),
           status: 404,
           headers: { 'Content-Type': 'application/json' }
         };
@@ -350,49 +351,6 @@ async function handleWatchApi(request, db, params, user) {
   };
 }
 
-async function handleTicketApi(request, env, params, user) {
-  const [action] = params;
-  
-  if (action === 'create' && request.method === 'POST') {
-    const { contentId, type, season, episode } = await request.json();
-    
-    // Generate unique ticket
-    const ticketId = generateTicketId();
-    const expiresAt = Date.now() + (6 * 60 * 60 * 1000); // 6 hours
-    
-    const ticketData = {
-      contentId,
-      type,
-      season,
-      episode,
-      userId: user?.id || 'guest',
-      expiresAt,
-      createdAt: Date.now()
-    };
-    
-    // Store in KV
-    await env.TICKETS.put(ticketId, JSON.stringify(ticketData), {
-      expirationTtl: 6 * 60 * 60 // 6 hours in seconds
-    });
-    
-    return {
-      body: JSON.stringify({
-        ticket: ticketId,
-        expiresAt,
-        streamUrl: `/stream/${ticketId}`
-      }),
-      status: 200,
-      headers: { 'Content-Type': 'application/json' }
-    };
-  }
-
-  return {
-    body: JSON.stringify({ error: 'Invalid request' }),
-    status: 400,
-    headers: { 'Content-Type': 'application/json' }
-  };
-}
-
 // Helper functions
 function formatMovie(movie) {
   return {
@@ -425,10 +383,4 @@ function formatTVShow(show) {
     genres: show.details?.genres || [],
     language: show.language
   };
-}
-
-function generateTicketId() {
-  return Array.from(crypto.getRandomValues(new Uint8Array(16)))
-    .map(b => b.toString(16).padStart(2, '0'))
-    .join('');
 }
